fix(box2d): scale collision tolerances with the length unit
The package exposes SetLengthUnitsPerMeter, but the tolerances upstream
derives from it were frozen Go constants, so a caller working in
non-meter units silently got upstream-divergent collision, continuous
collision and broad-phase behavior. Upstream evaluates them as macros
at use time.

Convert the length-scaled quantities -- LinearSlop, Huge,
SpeculativeDistance, ContactRecycleDistance and MaxAABBMargin -- to
variables recomputed by the setter, and leave the dimensionless
constants alone. Each default is derived from the same untyped constant
expression as before and only multiplied by a length unit of 1.0, which
is exact, so the defaults stay bit-identical; a test pins that with
Float64bits and the goldens are untouched.

Because LinearSlop is no longer constant-foldable, several expressions
that the compiler used to evaluate at compile time became runtime
products that could fuse into an FMA. TestNoFusedMultiplyAdd caught
them and they are rounded at their definitions.

Reported by an external review of this PR.

+++ pkg/box2d/distance.go
@@ -589,7 +589,7 @@ func ShapeCast(input *ShapeCastPairInput) CastOutput {
 	linearSlop := LinearSlop
 	totalRadius := input.ProxyA.Radius + input.ProxyB.Radius
 	target := maxFloat(linearSlop, totalRadius-linearSlop)
-	tolerance := 0.25 * linearSlop
+	tolerance := float64(0.25 * linearSlop)
 
 	assert(target > tolerance)
 
@@ -891,7 +891,7 @@ func TimeOfImpact(input *TOIInput) TOIOutput {
 	// Setup target distance and tolerance
 	totalRadius := proxyA.Radius + proxyB.Radius
 	target := maxFloat(LinearSlop, totalRadius-LinearSlop)
-	tolerance := 0.25 * LinearSlop
+	tolerance := float64(0.25 * LinearSlop)
 	assert(target > tolerance)
 
 	t1 := 0.0
